requirements.txt
django
django-taggit
django-autoslug
django-crispy-forms
crispy-bootstrap5